# swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "Gafarsa logistics api documentations",
    "version": "1.0.0",
    "title": "Gafarsa Logistics API Documentations",
    "termsOfService": "http://swagger.io/terms/",
    "contact": {
      "email": "apiteam@swagger.io"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "basePath": "/api/v1",
...
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "Enter JWT Bearer token as **Bearer &lt;token>**"
    }
  },
  "paths": {
...
    "/store/{id}": {
      "get": {
        "tags": ["Stores"],
        "summary": "Get store by ID",
        "description": "Retrieves a single store by ID.",
        "produces": ["application/json"],
        "security": [
          {
            "Bearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID of the store to retrieve",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "A single store.",
            "schema": {
              "$ref": "#/definitions/Store"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "User not found"
          }
        }
      },
      "put": {
        "tags": ["Stores"],
        "summary": "Update store by ID",
        "description": "Updates a single store's information.",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "security": [
          {
            "Bearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID of the store to update",
            "type": "string"
          },
          {
            "in": "body",
            "name": "store",
            "description": "User object with updated fields",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Store"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Store updated successfully."
          },
          "400": {
            "description": "Invalid Store supplied"
          },
          "404": {
            "description": "Store not found"
          }
        }
      },
      "delete": {
        "tags": ["Stores"],
        "summary": "Delete store by ID",
        "description": "Deletes a single store.",
        "security": [
          {
            "Bearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID of the store to delete",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Store deleted successfully."
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Store not found"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "Store": {
      "type": "object",
      "required": ["store_name", "category", "location", "logo"],
      "properties": {
        "store_name": {
          "type": "string"
        },
        "category": {
          "type": "string"
        },
        "sub_category": {
          "type": "string"
        },
        "address": {
          "type": "object",
          "properties": {
            "street": {
              "type": "string"
            },
            "city": {
              "type": "string"
            },
            "country": {
              "type": "string"
            },
            "postal_code": {
              "type": "string"
            }
          }
        },
        "website_link": {
          "type": "string"
        },
        "social_media_link": {
          "type": "string"
        },
        "business_hours": {
          "type": "string"
        },
        "location": {
          "type": "object",
          "properties": {
            "coordinates": {
              "type": "array",
              "items": {
                "type": "number"
              }
            },
            "type": {
              "type": "string"
            }
          }
        },
        "documents": {
          "type": "string"
        },
        "logo": {
          "type": "string"
        }
      }
    },
...
  }
}